tests: split tests and run template tests under both context behavior modes
