+++ tests/test_command_startcomponent.py
