+++ tests/test_component_media.py
@@ -1,0 +1,387 @@
+import sys
+from pathlib import Path
+
+from django.template import Context, Template
+from django.test import override_settings
+
+# isort: off
+from .django_test_setup import *  # NOQA
+from .testutils import BaseTestCase, autodiscover_with_cleanup
+
+# isort: on
+
+from django_components import component, types
+
+
+class InlineComponentTest(BaseTestCase):
+    def test_html(self):
+        class InlineHTMLComponent(component.Component):
+            template = "<div class='inline'>Hello Inline</div>"
+
+        comp = InlineHTMLComponent("inline_html_component")
+        self.assertHTMLEqual(
+            comp.render(Context({})),
+            "<div class='inline'>Hello Inline</div>",
+        )
+
+    def test_html_and_css(self):
+        class HTMLCSSComponent(component.Component):
+            template = "<div class='html-css-only'>Content</div>"
+            css = ".html-css-only { color: blue; }"
+
+        comp = HTMLCSSComponent("html_css_component")
+        self.assertHTMLEqual(
+            comp.render(Context({})),
+            "<div class='html-css-only'>Content</div>",
+        )
+        self.assertHTMLEqual(
+            comp.render_css_dependencies(),
+            "<style>.html-css-only { color: blue; }</style>",
+        )
+
+    def test_html_and_js(self):
+        class HTMLJSComponent(component.Component):
+            template = "<div class='html-js-only'>Content</div>"
+            js = "console.log('HTML and JS only');"
+
+        comp = HTMLJSComponent("html_js_component")
+        self.assertHTMLEqual(
+            comp.render(Context({})),
+            "<div class='html-js-only'>Content</div>",
+        )
+        self.assertHTMLEqual(
+            comp.render_js_dependencies(),
+            "<script>console.log('HTML and JS only');</script>",
+        )
+
+    def test_html_inline_and_css_js_files(self):
+        class HTMLStringFileCSSJSComponent(component.Component):
+            template = "<div class='html-string-file'>Content</div>"
+
+            class Media:
+                css = "path/to/style.css"
+                js = "path/to/script.js"
+
+        comp = HTMLStringFileCSSJSComponent("html_string_file_css_js_component")
+        self.assertHTMLEqual(
+            comp.render(Context({})),
+            "<div class='html-string-file'>Content</div>",
+        )
+        self.assertHTMLEqual(
+            comp.render_dependencies(),
+            """
+            <link href="path/to/style.css" media="all" rel="stylesheet">
+            <script src="path/to/script.js"></script>
+            """,
+        )
+
+    def test_html_js_inline_and_css_file(self):
+        class HTMLStringFileCSSJSComponent(component.Component):
+            template = "<div class='html-string-file'>Content</div>"
+            js = "console.log('HTML and JS only');"
+
+            class Media:
+                css = "path/to/style.css"
+
+        comp = HTMLStringFileCSSJSComponent("html_string_file_css_js_component")
+        self.assertHTMLEqual(
+            comp.render(Context({})),
+            "<div class='html-string-file'>Content</div>",
+        )
+        self.assertHTMLEqual(
+            comp.render_dependencies(),
+            """
+            <link href="path/to/style.css" media="all" rel="stylesheet">
+            <script>console.log('HTML and JS only');</script>
+            """,
+        )
+
+    def test_html_css_inline_and_js_file(self):
+        class HTMLStringFileCSSJSComponent(component.Component):
+            template = "<div class='html-string-file'>Content</div>"
+            css = ".html-string-file { color: blue; }"
+
+            class Media:
+                js = "path/to/script.js"
+
+        comp = HTMLStringFileCSSJSComponent("html_string_file_css_js_component")
+        self.assertHTMLEqual(
+            comp.render(Context({})),
+            "<div class='html-string-file'>Content</div>",
+        )
+        self.assertHTMLEqual(
+            comp.render_dependencies(),
+            """
+            <style>.html-string-file { color: blue; }</style><script src="path/to/script.js"></script>
+            """,
+        )
+
+    def test_html_variable(self):
+        class VariableHTMLComponent(component.Component):
+            def get_template(self, context):
+                return Template("<div class='variable-html'>{{ variable }}</div>")
+
+        comp = VariableHTMLComponent("variable_html_component")
+        context = Context({"variable": "Dynamic Content"})
+        self.assertHTMLEqual(
+            comp.render(context),
+            "<div class='variable-html'>Dynamic Content</div>",
+        )
+
+    def test_html_variable_filtered(self):
+        class FilteredComponent(component.Component):
+            template: types.django_html = """
+                Var1: <strong>{{ var1 }}</strong>
+                Var2 (uppercased): <strong>{{ var2|upper }}</strong>
+            """
+
+            def get_context_data(self, var1=None, var2=None):
+                return {
+                    "var1": var1,
+                    "var2": var2,
+                }
+
+        comp = FilteredComponent("filtered_component")
+        context = Context(comp.get_context_data(var1="test1", var2="test2"))
+
+        self.assertHTMLEqual(
+            comp.render(context),
+            """
+            Var1: <strong>test1</strong>
+            Var2 (uppercased): <strong>TEST2</strong>
+            """,
+        )
+
+
+class ComponentMediaTests(BaseTestCase):
+    def test_css_and_js(self):
+        class SimpleComponent(component.Component):
+            template: types.django_html = """
+                Variable: <strong>{{ variable }}</strong>
+            """
+
+            class Media:
+                css = "style.css"
+                js = "script.js"
+
+        comp = SimpleComponent("simple_component")
+        self.assertHTMLEqual(
+            comp.render_dependencies(),
+            """
+                <link href="style.css" media="all" rel="stylesheet">
+                <script src="script.js"></script>
+            """,
+        )
+
+    def test_css_only(self):
+        class SimpleComponent(component.Component):
+            template: types.django_html = """
+                Variable: <strong>{{ variable }}</strong>
+            """
+
+            class Media:
+                css = "style.css"
+
+        comp = SimpleComponent("simple_component")
+
+        self.assertHTMLEqual(
+            comp.render_dependencies(),
+            """
+            <link href="style.css" media="all" rel="stylesheet">
+            """,
+        )
+
+    def test_js_only(self):
+        class SimpleComponent(component.Component):
+            template: types.django_html = """
+                Variable: <strong>{{ variable }}</strong>
+            """
+
+            class Media:
+                js = "script.js"
+
+        comp = SimpleComponent("simple_component")
+
+        self.assertHTMLEqual(
+            comp.render_dependencies(),
+            """
+            <script src="script.js"></script>
+            """,
+        )
+
+    def test_empty_media(self):
+        class SimpleComponent(component.Component):
+            template: types.django_html = """
+                Variable: <strong>{{ variable }}</strong>
+            """
+
+            class Media:
+                pass
+
+        comp = SimpleComponent("simple_component")
+
+        self.assertHTMLEqual(comp.render_dependencies(), "")
+
+    def test_missing_media(self):
+        class SimpleComponent(component.Component):
+            template: types.django_html = """
+                Variable: <strong>{{ variable }}</strong>
+            """
+
+        comp = SimpleComponent("simple_component")
+
+        self.assertHTMLEqual(comp.render_dependencies(), "")
+
+    def test_css_js_as_lists(self):
+        class SimpleComponent(component.Component):
+            class Media:
+                css = ["path/to/style.css", "path/to/style2.css"]
+                js = ["path/to/script.js"]
+
+        comp = SimpleComponent("")
+        self.assertHTMLEqual(
+            comp.render_dependencies(),
+            """
+            <link href="path/to/style.css" media="all" rel="stylesheet">
+            <link href="path/to/style2.css" media="all" rel="stylesheet">
+            <script src="path/to/script.js"></script>
+            """,
+        )
+
+    def test_css_js_as_string(self):
+        class SimpleComponent(component.Component):
+            class Media:
+                css = "path/to/style.css"
+                js = "path/to/script.js"
+
+        comp = SimpleComponent("")
+        self.assertHTMLEqual(
+            comp.render_dependencies(),
+            """
+            <link href="path/to/style.css" media="all" rel="stylesheet">
+            <script src="path/to/script.js"></script>
+            """,
+        )
+
+    def test_css_js_as_dict_and_list(self):
+        class SimpleComponent(component.Component):
+            class Media:
+                css = {
+                    "all": "path/to/style.css",
+                    "print": ["path/to/style2.css"],
+                    "screen": "path/to/style3.css",
+                }
+                js = ["path/to/script.js"]
+
+        comp = SimpleComponent("")
+        self.assertHTMLEqual(
+            comp.render_dependencies(),
+            """
+            <link href="path/to/style.css" media="all" rel="stylesheet">
+            <link href="path/to/style2.css" media="print" rel="stylesheet">
+            <link href="path/to/style3.css" media="screen" rel="stylesheet">
+            <script src="path/to/script.js"></script>
+            """,
+        )
+
+class MediaRelativePathTests(BaseTestCase):
+    class ParentComponent(component.Component):
+        template: types.django_html = """
+            {% load component_tags %}
+            <div>
+                <h1>Parent content</h1>
+                {% component name="variable_display" shadowing_variable='override' new_variable='unique_val' %}
+                {% endcomponent %}
+            </div>
+            <div>
+                {% slot 'content' %}
+                    <h2>Slot content</h2>
+                    {% component name="variable_display" shadowing_variable='slot_default_override' new_variable='slot_default_unique' %}
+                    {% endcomponent %}
+                {% endslot %}
+            </div>
+        """  # noqa
+
+        def get_context_data(self):
+            return {"shadowing_variable": "NOT SHADOWED"}
+
+    class VariableDisplay(component.Component):
+        template: types.django_html = """
+            {% load component_tags %}
+            <h1>Shadowing variable = {{ shadowing_variable }}</h1>
+            <h1>Uniquely named variable = {{ unique_variable }}</h1>
+        """
+
+        def get_context_data(self, shadowing_variable=None, new_variable=None):
+            context = {}
+            if shadowing_variable is not None:
+                context["shadowing_variable"] = shadowing_variable
+            if new_variable is not None:
+                context["unique_variable"] = new_variable
+            return context
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        component.registry.register(name="parent_component", component=cls.ParentComponent)
+        component.registry.register(name="variable_display", component=cls.VariableDisplay)
+
+    # Settings required for autodiscover to work
+    @override_settings(
+        BASE_DIR=Path(__file__).resolve().parent,
+        STATICFILES_DIRS=[
+            Path(__file__).resolve().parent / "components",
+        ],
+    )
+    def test_component_with_relative_media_paths(self):
+        # Ensure that the module is executed again after import in autodiscovery
+        if "tests.components.relative_file.relative_file" in sys.modules:
+            del sys.modules["tests.components.relative_file.relative_file"]
+
+        # Fix the paths, since the "components" dir is nested
+        with autodiscover_with_cleanup(map_import_paths=lambda p: f"tests.{p}"):
+            template_str: types.django_html = """
+                {% load component_tags %}{% component_dependencies %}
+                {% component name='relative_file_component' variable=variable %}
+                {% endcomponent %}
+            """
+            template = Template(template_str)
+            rendered = template.render(Context({"variable": "test"}))
+            self.assertHTMLEqual(
+                rendered,
+                """
+                <link href="relative_file/relative_file.css" media="all" rel="stylesheet">
+                <script src="relative_file/relative_file.js"></script>
+                <form method="post">
+                    <input type="text" name="variable" value="test">
+                    <input type="submit">
+                </form>
+                """,
+            )
+
+    # Settings required for autodiscover to work
+    @override_settings(
+        BASE_DIR=Path(__file__).resolve().parent,
+        STATICFILES_DIRS=[
+            Path(__file__).resolve().parent / "components",
+        ],
+    )
+    def test_component_with_relative_media_paths_as_subcomponent(self):
+        # Ensure that the module is executed again after import in autodiscovery
+        if "tests.components.relative_file.relative_file" in sys.modules:
+            del sys.modules["tests.components.relative_file.relative_file"]
+
+        # Fix the paths, since the "components" dir is nested
+        with autodiscover_with_cleanup(map_import_paths=lambda p: f"tests.{p}"):
+            template_str: types.django_html = """
+                {% load component_tags %}{% component_dependencies %}
+                {% component 'parent_component' %}
+                    {% fill 'content' %}
+                        {% component name='relative_file_component' variable='hello' %}
+                        {% endcomponent %}
+                    {% endfill %}
+                {% endcomponent %}
+            """
+            template = Template(template_str)
+            rendered = template.render(Context({}))
+            self.assertIn('<input type="text" name="variable" value="hello">', rendered, rendered)
